+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.16.4] Q4 2023
+- ssh tunnel:
+  - set 'https://cp.omnect.conplement.cloud' as default backend
+  - updated ssh tunnel readme
+- improved error handling and output
+
 ## [0.16.3] Q4 2023
 - added copy from image command
 - removed dedicated wifi command but added example with copy-to-image command in README